// client-id/src/app/core/interfaces/form-consent-complete-rp.ts
import { ConsentConfmProcessHist } from './form-consent-complete-view';

/* eslint-disable @typescript-eslint/naming-convention */
export interface FormConsentCompleteRp{
    gridData: formConsentData[];
}

export interface formConsentData{
    fomConfmSn: number;
    beforeFomConfmSn: number;
    frstFomConfmSn: number;
    fomConfmProcessSttusCode: string;
    fomConfmProcessSttusName: string;
    fomConfmProcessSttusName_User: string;
    proprod: number;
    trnsfrEntrpsSn: number;
    entrpsSn: number;
    entrprsNm: string;
    rprsntvNm: string;
    bizrno: string;
    bplcAdresSeCode: string;
    bplcZip: string;
    bplcBassAdres: string ;
    bplcDetailAdres:   string;
    bplcTelno: string;
    bplcFxnum: string;
    fctryAdresSeZip: string;
    fctryZip: string;
    fctryBassAdres: string;
    fctryDetailAdres: string;
    fctryTelno: string;
    fctryFxnum: string;
    mrnrSeCode: string;
    modelNm: string;
    registSeCode: string;
    confmReqstdocCn: string;
    confmReqstdocFeeCn: string;
    testFeesm: number;
    bsrpTrnsportCt: number;
    bsrpDeCtDaycnt: number;
    bsrpDeCtAmount: number;
    bsrpCgffdAmount: number;
    bsrpStayngCtDaycnt: number;
    bsrpStayngCtAmount: number;
    btrpsSm: number;
    fomConfmIsgnCt: number;
    subsum: number;
    vat: number;
    totalSmAm: number;
    payAt: string;
    feePayDe: Date;
    feePayCnfrmrId: string;
    feePayCnfrmrNm: string;
    feePayCnfrmrDeptNm: string;
    sploreQy: number;
    testSploreProcessCode: string;
    testSploreProcessCodeName: string;
    sploreRceptDe: Date;
    sploreChargerId: string;
    sploreChargerNm: string;
    sploreChargerDeptNm: string;
    testResultPsexamAt: string;
    testComptDe: Date;
    fomConfmInsttSeCode: string;
    changeReqstDtls: string;
    changeProcessDtls: string;
    fomComTcyStNdNtNo: string;
    fomConfmProcessCn: string;
    fomConfmReqstNo: number;
    fomConfmRceptNo: string;
    fomConfmConfmNo: string;
    fomConfmSeCode: string;
    etcMatterCn: string;
    isgnNbfps: number;
    rqstDt: Date;
    applcntId: string;
    applcntNm: string;
    applcntTelno: string;
    applcntMbtlnum: string;
    applcntEmailAdres: string;
    rceptDe: Date;
    rceptChargerId: string;
    rceptChargerNm: string;
    rceptChargerDeptNm: string;
    rceptConfmDt: Date;
    rceptConfmerId: string;
    rceptConfmerNm: string;
    rceptConfmerDeptNm: string;
    fomConfmChargerId: string;
    fomConfmChargerNm: string;
    fomConfmChargerDeptNm: string;
    fomConfmChargerTelno: string;
    lastConfmDe: Date;
    fomConfmRspnberId: string;
    fomConfmRspnberNm: string;
    fomConfmRspnberDeptNm: string;
    ceratsPblictePrarnDe: Date;
    taxbilPblicteAt: string;
    mnfctRegistKndCn: string;
    mnfcturRegistDe: string;
    mrnrSeName: string;
    fomConfmInsttSeName: string;
    registSeName: string;
    fomConfmProcessSttusCodeName: string;
    payAt_String: string;
    testResultPsexamAt_String: string;
    statusClassName: string;
    rowNumber: number;
    totalCount: number;
    isReIssueExist: string;
    fomConfmProcessHistList: ConsentConfmProcessHist[];
}